+++ docs/CHANGE_LOG.md
@@ -1,5 +1,12 @@
 # JHson.js - Change Log:
 
+## Version 1.1.0:
+- Added "use strict" support internally and updated all public functions to use the new scope.
+- Fixed a fault that caused the function parameter "properties" to collide with an outer scope variable of the same name for "json() > ignoreCssProperties()".
+
+<br>
+
+
 ## Version 1.0.0:
 
 #### **New Features:**